Consolidate test files and update dev snapshots
